// react-app-common/gen/schemas/v1/todo_connect.ts
// @generated by protoc-gen-connect-es v0.8.4 with parameter "target=ts,import_extension=none"
// @generated from file schemas/v1/todo.proto (package schemas.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { Empty, MethodKind } from "@bufbuild/protobuf";
import { CreateTodoRequest, CreateTodoResponse, DestroyTodoRequest, IndexTodoResponse, UpdateTodoRequest, UpdateTodoResponse } from "./todo_pb";

/**
 * Todo の各種 Service。
 * Rails の Actions にならった命名を行っている。
 * see: https://www.codecademy.com/article/standard-controller-actions
 *
 * @generated from service schemas.v1.TodoService
 */
export const TodoService = {
  typeName: "schemas.v1.TodoService",
  methods: {
    /**
     * Todo の一覧を取得する。
     *
     * @generated from rpc schemas.v1.TodoService.IndexTodo
     */
    indexTodo: {
      name: "IndexTodo",
      I: Empty,
      O: IndexTodoResponse,
      kind: MethodKind.Unary,
    },
    /**
     * Todo を新規追加する。
     *
     * @generated from rpc schemas.v1.TodoService.CreateTodo
     */
    createTodo: {
      name: "CreateTodo",
      I: CreateTodoRequest,
      O: CreateTodoResponse,
      kind: MethodKind.Unary,
    },
    /**
     * Todo を更新する。
     *
     * @generated from rpc schemas.v1.TodoService.UpdateTodo
     */
    updateTodo: {
      name: "UpdateTodo",
      I: UpdateTodoRequest,
      O: UpdateTodoResponse,
      kind: MethodKind.Unary,
    },
    /**
     * Todo を削除する。
     *
     * @generated from rpc schemas.v1.TodoService.DestroyTodo
     */
    destroyTodo: {
      name: "DestroyTodo",
      I: DestroyTodoRequest,
      O: Empty,
      kind: MethodKind.Unary,
    },
  }
} as const;